Add files via upload
The purpose of this project is to define  factors from the chosen dataset which can be used to predict a Maths exam mark.

Our dataset contains less than 400 records and around 300 of them have at least 1 column with missing value. Out of 28 columns which we potentially add to the final ML model, most contain descriptive categorical values. None of them have a significant p value to be certain to tell that these fields can definitely be used  in a  predictive model.

Overall, tehri is no clear view what should be included in the model and some regressions analysis need to be run to define the best possible combination of factors.